[chore] feature_job_setting_analysis: split backtest to separate file
